new file:   CutFlowTable.C 	new file:   CutFlowTable.h 	modified:   SmallClass.h 	new file:   clean.sh 	new file:   pre_run.C 	modified:   run.C 	modified:   run.sh
The new class CutFlowTable is defined. This class is in the files: CutFlowTable.C and CutFlowTable.h.

If new cut is applied in the analysis the a new method will be added to "SmallClassExtra.C" and "SmallClassExtra.h".
Then a new reasonable name for cut must be added in the file "CutFlowTable.h" for memebers called: "cut_name_string" and "cut_name". The "cut_name" is an enum.
The index of "cut_name" must be compatiable with "cut_name_string" index.

After that the new cut can be applied in "run.C" with "fill" method of CutFlowTable class.
CutFlowTable instance is initialized with number of entries in the "run.C" file.

For example for trigger cut in analysis we do this:

1- turn on branches in root file related to trigger (in "turn_on_necessary_branches" method of "SmallClassExtra" in the "SmallClassExtra.C" file)
2- define the cut method with name "trigger_cut()" in "SmallClass.h" and "SmallClass.C"
3- add the string "trigger" as 0 element of member "cut_name_string" in "CutFlowTable" class which is in "CutFlowTable.h" file.
4- add the name "trigger" to enum class called "cut_name" in the "CutFlowTable" class which is in the "CutFlowTable.h" file.
5- The value of "trigger" in "cut_name" shoud set to 0, because in the "cut_name_string" we have written "trigger" as 0th element.
6- Warning: Check 5 again. That can cause mistakes.
7- In "run.C" file, in event loop, apply the cut using "tigger_cut()" method, and update "cut_flow_table" object using "fill" method with the name of the cut:
fill(CutFlowTable::cut_name::trigger);

---------- pre_run.C -------------
a print functions and the files we shuold include are moved from "run.C" to "pre_run.C".

---------- run.sh ----------------
modified run.sh to  print cut flow table in the screen.

----------- clean.sh -------------
create clean.sh to delete outputs of the analysis code, which are in the *.txt or *.csv format.
Warning: Do not write a code file with *.txt or *.csv extension format.